# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Notification Service API
  description: |
    Comprehensive Notification-as-a-Service API for high-frequency multichannel delivery.
    Supports Email, SMS, Push, Webhooks, WebSockets, and Slack (Incoming Webhook–style payloads).

    **Slack:** use channel `slack`. Configure the Incoming Webhook URL in the App Store / vendor config
    (`slack` / `webhook_url`). Set `recipient` to that webhook URL (or your deployment's resolved address),
    and send message text via `body`, a template, or `template_variables` keys `message` / `text`.
    Async worker processing uses the Pub/Sub topic `notifications-slack`.

    **Authentication:** All endpoints require a Clerk session JWT `Bearer` token (via the Clerk SDK) unless noted otherwise.
    Legacy HMAC JWT tokens are accepted during the migration period.
    Service-to-service endpoints (OTP) use the `X-Service-Token` header instead.
    Programmatic access via `X-API-Key` is also supported.
  version: 1.1.0
  contact:
    name: Spidey Deepmind Team
    url: https://github.com/spidey/notification-service
servers:
  - url: http://localhost:8080/v1
    description: Local Development API (v1)
  - url: http://localhost:8080
    description: Local Development root endpoints
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    ServiceAuth:
      type: apiKey
      in: header
      name: X-Service-Token
  schemas:
    Channel:
      type: string
      enum: [email, sms, push, websocket, webhook, slack]
...
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /internal/circuit-breakers:
    get:
      summary: Circuit breaker snapshot (internal only)
      tags: [System]
      security: []
      servers:
        - url: http://localhost:8080
      responses:
        '200':
          description: Circuit breaker states
        '403':
          description: Access restricted to internal network
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── Auth ──────────────────────────────────────────────────────────────────
  # Authentication is handled by Clerk (https://clerk.com).
  # UI: Use Clerk's <SignIn /> component or redirect to your Clerk portal.
  # API: Include Clerk session JWT in the Authorization header.

  # Provider delivery webhooks/callbacks
  /webhooks/{provider}:
    parameters:
      - name: provider
        in: path
        required: true
        schema: { type: string }
    post:
      summary: Provider webhook callback (POST)
      description: Inbound provider delivery/status webhook. Signature validation is provider-specific.
      tags: [Webhooks]
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Invalid payload
        '401':
          description: Invalid signature
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: Provider webhook callback (GET)
      description: Some providers deliver status callbacks via GET query parameters.
      tags: [Webhooks]
      security: []
      responses:
        '200':
          description: Callback processed
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Clerk webhook for user sync
  /webhooks/clerk:
    post:
      summary: Clerk user lifecycle webhook
      tags: [Webhooks]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Clerk webhook payload (user.created, user.updated, user.deleted)
      responses:
        '200':
          description: Webhook processed
        '401':
          description: Invalid webhook signature
        '500':
          $ref: '#/components/responses/InternalServerError'

  # API self-documentation endpoints
  /openapi.yaml:
    get:
      summary: Get OpenAPI YAML document
      tags: [System]
      security: []
      responses:
        '200':
          description: OpenAPI YAML
        '500':
          $ref: '#/components/responses/InternalServerError'
  /openapi.json:
    get:
      summary: Get OpenAPI JSON document
      tags: [System]
      security: []
      responses:
        '200':
          description: OpenAPI JSON
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── Notifications ─────────────────────────────────────────────────────────
  /notifications:
    post:
      summary: Send a notification
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendRequest' }
      responses:
        '202':
          description: Notification accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification_id: { type: string, format: uuid }
                  status: { type: string }
                  workflow_id: { type: string }
                  scheduled_at: { type: string, format: date-time, nullable: true }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: List notifications
      tags: [Notifications]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: user_id
          in: query
          schema: { type: string, format: uuid }
        - name: channel
          in: query
          schema: { $ref: '#/components/schemas/Channel' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/NotificationStatus' }
        - name: recipient
          in: query
          description: Filter by recipient (partial match, case-insensitive).
          schema: { type: string }
        - name: search
          in: query
          description: Full-text search across recipient, notification ID, and idempotency key.
          schema: { type: string }
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/ApiKeyIdScope'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Notification' }
                  total: { type: integer }
                  page: { type: integer }
                  page_size: { type: integer }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /notifications/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get notification details (with attempts and events)
      tags: [Notifications]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification: { $ref: '#/components/schemas/Notification' }
                  attempts:
                    type: array
                    items: { $ref: '#/components/schemas/NotificationAttempt' }
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/NotificationEvent' }
                  subject: { type: string }
                  body: { type: string }
                  provider_status:
                    type: object
                    nullable: true
                    properties:
                      vendor_status: { type: string }
                      success: { type: boolean }
                      provider: { type: string }
                      provider_msg_id: { type: string }
                      error_code: { type: string }
                      error_message: { type: string }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /notifications/{id}/sync:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Sync delivery status from the vendor
      description: |
        Polls the originating provider for the latest delivery status and updates the notification
        record. Only applicable to notifications in non-final states (pending, queued, sent).
      tags: [Notifications]
      responses:
        '200':
          description: Synced
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification_id: { type: string, format: uuid }
                  success: { type: boolean }
                  vendor_status: { type: string }
                  provider: { type: string }
                  provider_msg_id: { type: string }
                  error_code: { type: string }
                  error_message: { type: string }
        '400':
          description: Invalid sync state (e.g. no provider message id)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /notifications/{id}/retrigger:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Retrigger delivery of a failed or stuck notification
      description: Starts a new delivery workflow for the existing notification. Does not create a new record.
      tags: [Notifications]
      responses:
        '200':
          description: Retriggered
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification_id: { type: string, format: uuid }
                  status: { type: string }
        '400':
          description: Retrigger failed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── Scheduled notifications ───────────────────────────────────────────────
  /notifications/scheduled:
    get:
      summary: List scheduled notifications
      tags: [Scheduled]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/NotificationStatus' }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ScheduledNotification' }
                  total: { type: integer }
                  page: { type: integer }
                  page_size: { type: integer }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /notifications/{id}/schedule:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Reschedule a pending notification
      tags: [Scheduled]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scheduled_at]
              properties:
                scheduled_at: { type: string, format: date-time, description: New dispatch time (must be in the future). }
      responses:
        '200':
          description: Rescheduled
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification_id: { type: string, format: uuid }
                  status: { type: string }
                  scheduled_at: { type: string, format: date-time }
                  workflow_id: { type: string }
                  reschedule_count: { type: integer }
        '400':
          description: Invalid time or not reschedulable
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Cancel a scheduled notification
      tags: [Scheduled]
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification_id: { type: string, format: uuid }
                  status: { type: string, example: cancelled }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── OTP ───────────────────────────────────────────────────────────────────
  /otp/send:
    post:
      summary: Generate and send OTP
      tags: [OTP]
      security:
        - ServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OTPSendRequest' }
      responses:
        '200':
          description: OTP Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  otp_id: { type: string }
                  expiry_at: { type: string, format: date-time }
        '400':
          description: Validation error
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /otp/verify:
    post:
      summary: Verify an OTP
      tags: [OTP]
      security:
        - ServiceAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OTPVerifyRequest' }
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified: { type: boolean }
        '400':
          description: Validation error or invalid OTP
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── Reports ───────────────────────────────────────────────────────────────
  /reports/summary:
    get:
      summary: Delivery summary by channel and date
      description: Returns aggregated sent/delivered/failed/bounced counts plus latency percentiles, grouped by channel and calendar day.
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/ApiKeyIdScope'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReportSummaryItem' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/ingress:
    get:
      summary: Ingress source breakdown
      description: Returns the count of notification requests grouped by ingress source (api, pubsub, redis).
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/ApiKeyIdScope'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IngressBreakdownItem' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/sms-countries:
    get:
      summary: SMS count by country prefix
      description: Returns SMS notification counts grouped by destination country (derived from the E.164 phone prefix). Top 10 countries.
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/ApiKeyIdScope'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BreakdownRow' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/email-domains:
    get:
      summary: Email count by recipient domain
      description: Returns email notification counts grouped by recipient domain (e.g. gmail.com). Top 10 domains.
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - $ref: '#/components/parameters/ApiKeyIdScope'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/BreakdownRow' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/vendors:
    get:
      summary: Real-time vendor health metrics
      description: |
        Returns per-provider delivery metrics computed over the last 12 hours.
        Refreshed on each call; callers should poll at a reasonable interval (e.g. 10 s).
      tags: [Reports]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/VendorMetric' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /reports/billing:
    get:
      summary: Vendor billing and cost summary
      description: |
        Returns per-vendor cost data. For providers with a billing API (Twilio, Plivo, Vonage,
        MessageBird), the cost is fetched live. For others (SES, SendGrid, Postmark, Mailgun),
        cost is estimated from DB send counts × published per-message rate.
        Free-tier providers (FCM, Slack, Webhook) show zero cost.
        Response is cached for 5 minutes on the client; server calls vendor APIs with a 12-second timeout.
      tags: [Reports]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/VendorBilling' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /governance/suppressions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Remove a suppression
      description: Removes the suppression entry. Future notifications to this identifier will be delivered normally.
      tags: [Governance]
      responses:
        '204':
          description: Removed
        '400':
          description: Invalid ID
        '401':
          description: Unauthorized
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...